experiments/mvp-adk-agent-with-ca/requirements.txt
google-adk>=1.3.0
google-cloud-geminidataanalytics>=0.10.0
